Improve example comments
