--- substrate-airgapped/src/transaction/mod.rs
@@ -1,8 +1,0 @@
-pub(crate) mod extra;
-mod generic_call;
-mod tx;
-
-pub use self::{
-	generic_call::{CallIndex, GenericCall},
-	tx::{tx_from_parts, MortalConfig, Mortality, SignedPayload, Tx, TxConfig, UncheckedExtrinsic},
-};
