+++ substrate-airgapped/src/tx/mod.rs
@@ -1,7 +1,14 @@
-use super::{
-	extra::{Extra, SignedExtra},
-	GenericCall,
+pub(crate) mod extra;
+mod generic_call;
+mod mortality;
+
+pub use self::{
+	generic_call::{CallIndex, GenericCall},
+	// tx::{tx_from_parts, SignedPayload, Tx, TxConfig, UncheckedExtrinsic},
+	mortality::{MortalConfig, Mortality}
 };
+
+use self::extra::{Extra, SignedExtra};
 use crate::{
 	frame::{balances::Balances, system::System},
 	runtimes::Runtime,
@@ -21,29 +28,6 @@ pub type UncheckedExtrinsic<C, R> = sp_runtime::generic::UncheckedExtrinsic<
 /// Local `SignedPayload` convenience type. This is the payload that gets signed.
 pub type SignedPayload<C, R> = sp_runtime::generic::SignedPayload<GenericCall<C>, Extra<R>>;
 
-/// Mortal period configuration options,
-///
-/// Read here for conceptual details: https://docs.rs/sp-runtime/2.0.0/sp_runtime/generic/enum.Era.html
-#[derive(Clone, Eq, PartialEq, Debug, Copy)]
-pub struct MortalConfig<R: System> {
-	/// Duration of the transactions validity, measured in blocks, starting from the checkpoint block.
-	pub period: u64,
-	/// Hash of the block where the transaction's mortality period starts.
-	pub checkpoint_block_hash: R::Hash,
-	/// Block number where the transaction mortality period starts.
-	pub checkpoint_block_number: u64,
-}
-/// Specify the mortality of a transaction.
-///
-/// Read here for conceptual details: https://docs.rs/sp-runtime/2.0.0/sp_runtime/generic/enum.Era.html
-#[derive(Clone, Eq, PartialEq, Debug, Copy)]
-pub enum Mortality<R: System> {
-	/// Specify a mortal transaction with period, checkpoint block number, and
-	/// checkpoint block hash
-	Mortal(MortalConfig<R>),
-	/// Specify an immortal transaction
-	Immortal,
-}
 /// Configuration options for a Tx
 #[derive(Clone, PartialEq, Debug)]
 pub struct TxConfig<C: Encode + Decode + Clone, R: System + Balances + Runtime> {
